media_processor/media_processor/requirements.txt
langgraph>=0.2.0
opencv-python>=4.8.0.0
ffmpeg-python>=0.2.0
